Extract ShimmerRowWithIcon as a component
