+++ app/src/main/java/com/jp_funda/todomind/view/mind_map/components/DummyMindMapsRow.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
-import com.jp_funda.todomind.view.components.AnimatedShimmer
+import com.jp_funda.todomind.view.components.shimmer.AnimatedShimmer
 
 @Composable
 fun DummyMindMapsRow() {
