+++ app/src/main/java/com/jp_funda/todomind/view/mind_map_detail/components/MindMapDetailLoadingContent.kt
@@ -1,0 +1,8 @@
+package com.jp_funda.todomind.view.mind_map_detail.components
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun MindMapDetailLoadingContent() {
+
+}
