+++ app/src/main/java/com/jp_funda/todomind/view/task_detail/components/TaskEditLoadingContent.kt
@@ -1,14 +1,13 @@
 package com.jp_funda.todomind.view.task_detail.components
 
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
-import com.jp_funda.todomind.view.components.AnimatedShimmer
+import com.jp_funda.todomind.view.components.shimmer.AnimatedShimmer
+import com.jp_funda.todomind.view.components.shimmer.ShimmerRowWithIcon
 
 @Composable
 fun TaskEditLoadingContent() {
@@ -63,26 +62,5 @@ fun TaskEditLoadingContent() {
                     .fillMaxWidth(0.5f)
             )
         }
-    }
-}
-
-@Composable
-fun ShimmerRowWithIcon(modifier: Modifier) {
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        AnimatedShimmer(
-            modifier = Modifier
-                .size(50.dp)
-                .clip(CircleShape)
-        )
-        Spacer(modifier = Modifier.width(10.dp))
-        AnimatedShimmer(
-            modifier = Modifier
-                .height(50.dp)
-                .clip(RoundedCornerShape(10.dp))
-                .fillMaxWidth()
-        )
     }
 }
